# Backend/Swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user": {
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "post": {
        "operationId": "UserController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSchemaDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateSchemaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
...
    "/chat/fetch/user/{userId}": {
      "get": {
        "operationId": "ChatController_getChatRoomsByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Chat"
        ]
      }
    },
...
    "/support": {
      "get": {
        "operationId": "SupportController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Support"
        ]
      },
      "post": {
        "operationId": "SupportController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSupportSchemaDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Support"
        ]
      }
    },
    "/support/user": {
      "get": {
        "operationId": "SupportController_getAllByUserId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Support Ticket Id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Support"
        ]
      }
    },
    "/support/{id}": {
      "get": {
        "operationId": "SupportController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Support Ticket Id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Support"
        ]
      },
      "patch": {
        "operationId": "SupportController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Support Ticket Id",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSupportSchemaDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Support"
        ]
      },
      "delete": {
        "operationId": "SupportController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Support"
        ]
      }
    }
  },
  "info": {
    "title": "ExBuySell - Api Documentation",
    "description": "",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "UserSchemaDTO": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "minLength": 4
          },
          "last_name": {
            "type": "string",
            "minLength": 4
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER",
              "MONITER"
            ]
          },
          "businessName": {
            "type": "string",
            "nullable": true
          },
          "contactName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "permissions": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "zip_code": {
            "type": "integer",
            "format": "int64"
          },
          "background": {
            "type": "string"
          },
          "is_online": {
            "default": false,
            "type": "boolean"
          },
          "password_hash": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string",
            "nullable": true
          },
          "otp_code": {
            "type": "string"
          },
          "approved": {
            "default": false,
            "type": "boolean"
          },
          "is_email_verified": {
            "default": false,
            "type": "boolean"
          },
          "is_phone_verified": {
            "default": false,
            "type": "boolean"
          }
        },
        "required": [
          "first_name",
          "last_name",
          "role",
          "contactName",
          "email",
          "background",
          "password_hash"
        ]
      },
      "UserUpdateSchemaDTO": {
        "type": "object",
        "properties": {
          "first_name": {
            "type": "string",
            "minLength": 4
          },
          "last_name": {
            "type": "string",
            "minLength": 4
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "USER",
              "MONITER"
            ]
          },
          "businessName": {
            "type": "string",
            "nullable": true
          },
          "contactName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "permissions": {
            "items": {
              "type": "array"
            },
            "type": "array"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "zip_code": {
            "type": "integer",
            "format": "int64"
          },
          "background": {
            "type": "string"
          },
          "is_online": {
            "default": false,
            "type": "boolean"
          },
          "password_hash": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string",
            "nullable": true
          },
          "otp_code": {
            "type": "string"
          },
          "approved": {
            "default": false,
            "type": "boolean"
          },
          "is_email_verified": {
            "default": false,
            "type": "boolean"
          },
          "is_phone_verified": {
            "default": false,
            "type": "boolean"
          }
        }
      },
...
      "CreateSupportSchemaDTO": {
        "type": "object",
        "properties": {
          "chatRoomId": {
            "type": "string",
            "minLength": 4
          },
          "message": {
            "type": "string",
            "minLength": 4
          },
          "subject": {
            "type": "string",
            "minLength": 4
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "CLOSED",
              "PENDING"
            ]
          }
        },
        "required": [
          "chatRoomId",
          "message",
          "subject"
        ]
      },
      "UpdateSupportSchemaDTO": {
        "type": "object",
        "properties": {
          "chatRoomId": {
            "type": "string",
            "minLength": 4
          },
          "message": {
            "type": "string",
            "minLength": 4
          },
          "subject": {
            "type": "string",
            "minLength": 4
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "CLOSED",
              "PENDING"
            ]
          }
        }
      }
    }
  }
}